+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/createplan/CreatePlanScreen.kt
@@ -1,25 +1,25 @@
-package com.yapp.growth.presentation.ui.home
+package com.yapp.growth.presentation.ui.main.createplan
 
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 
 @Composable
-fun HomeScreen() {
+fun CreatePlanScreen() {
     Surface(
         modifier = Modifier.fillMaxSize(),
-        color = Color(0xFFF7F7F8)
+        color = MaterialTheme.colors.background
     ) {
-        Text(text = "Home")
+        Text(text = "Create Plan")
     }
 }
 
 @Preview
 @Composable
 fun HomeScreenPreview() {
-    HomeScreen()
+    CreatePlanScreen()
 }
