+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/home/HomeScreen.kt
@@ -1,25 +1,25 @@
-package com.yapp.growth.presentation.ui.manageplan
+package com.yapp.growth.presentation.ui.main.home
 
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 
 @Composable
-fun ManagePlanScreen() {
+fun HomeScreen() {
     Surface(
         modifier = Modifier.fillMaxSize(),
-        color = MaterialTheme.colors.background
+        color = Color(0xFFF7F7F8)
     ) {
-        Text(text = "Manage Plan")
+        Text(text = "Home")
     }
 }
 
 @Preview
 @Composable
 fun HomeScreenPreview() {
-    ManagePlanScreen()
+    HomeScreen()
 }
